refactor(mypage): 마이페이지 전용 컴포넌트 위치 정리
