+++ src/pages/mypage/components/FavoriteGameCardSkeleton.tsx
