+++ src/pages/mypage/components/FavoriteGamesEmptyState.tsx
